+++ inquire/src/prompts/multiselect/mod.rs
@@ -16,13 +16,19 @@ use crate::{
     list_option::ListOption,
     prompts::prompt::Prompt,
     terminal::get_default_terminal,
-    type_aliases::Filter,
+    type_aliases::Scorer,
     ui::{Backend, MultiSelectBackend, RenderConfig},
     validator::MultiOptionValidator,
 };
 
 use self::prompt::MultiSelectPrompt;
 
+#[cfg(feature = "fuzzy")]
+use fuzzy_matcher::{skim::SkimMatcherV2, FuzzyMatcher};
+#[cfg(feature = "fuzzy")]
+use once_cell::sync::Lazy;
+#[cfg(feature = "fuzzy")]
+static DEFAULT_MATCHER: Lazy<SkimMatcherV2> = Lazy::new(|| SkimMatcherV2::default().ignore_case());
 /// Prompt suitable for when you need the user to select many options (including none if applicable) among a list of them.
 ///
 /// The user can select (or deselect) the current highlighted option by pressing space, clean all selections by pressing the left arrow and select all options by pressing the right arrow.
@@ -38,6 +44,7 @@ use self::prompt::MultiSelectPrompt;
 /// - **Options list**: Options displayed to the user. Must be **non-empty**.
 /// - **Default selections**: Options that are selected by default when the prompt is first rendered. The user can unselect them. If any of the indices is out-of-range of the option list, the prompt will fail with an [`InquireError::InvalidConfiguration`] error.
 /// - **Starting cursor**: Index of the cursor when the prompt is first rendered. Default is 0 (first option). If the index is out-of-range of the option list, the prompt will fail with an [`InquireError::InvalidConfiguration`] error.
+/// - **Starting filter input**: Sets the initial value of the filter section of the prompt.
 /// - **Help message**: Message displayed at the line below the prompt.
 /// - **Formatter**: Custom formatter in case you need to pre-process the user input before showing it as the final answer.
 ///   - Prints the selected options string value, joined using a comma as the separator, by default.
@@ -45,7 +52,7 @@ use self::prompt::MultiSelectPrompt;
 ///   - No validators are on by default.
 /// - **Page size**: Number of options displayed at once, 7 by default.
 /// - **Display option indexes**: On long lists, it might be helpful to display the indexes of the options to the user. Via the `RenderConfig`, you can set the display mode of the indexes as a prefix of an option. The default configuration is `None`, to not render any index when displaying the options.
-/// - **Filter function**: Function that defines if an option is displayed or not based on the current filter input.
+/// - **Scorer function**: Function that defines the order of options and if displayed as all.
 /// - **Keep filter flag**: Whether the current filter input should be cleared or not after a selection is made. Defaults to true.
 ///
 /// # Example
@@ -77,9 +84,17 @@ pub struct MultiSelect<'a, T> {
     /// Starting cursor index of the selection.
     pub starting_cursor: usize,
 
-    /// Function called with the current user input to filter the provided
+    /// Starting filter input
+    pub starting_filter_input: Option<&'a str>,
+
+    /// Reset cursor position to first option on filter input change.
+    /// Defaults to true.
+    pub reset_cursor: bool,
+
+    /// Function called with the current user input to score the provided
     /// options.
-    pub filter: Filter<'a, T>,
+    /// The list of options is sorted in descending order (highest score first)
+    pub scorer: Scorer<'a, T>,
 
     /// Whether the current filter typed by the user is kept or cleaned after a selection is made.
     pub keep_filter: bool,
@@ -134,34 +149,44 @@ where
             .join(", ")
     };
 
-    /// Default filter function, which checks if the current filter value is a substring of the option value.
-    /// If it is, the option is displayed.
+    /// Default scoring function, which will create a score for the current option using the input value.
+    /// The return will be sorted in Descending order, leaving options with None as a score.
     ///
     /// # Examples
     ///
     /// ```
     /// use inquire::MultiSelect;
     ///
-    /// let filter = MultiSelect::<&str>::DEFAULT_FILTER;
-    /// assert_eq!(false, filter("sa", &"New York",      "New York",      0));
-    /// assert_eq!(true,  filter("sa", &"Sacramento",    "Sacramento",    1));
-    /// assert_eq!(true,  filter("sa", &"Kansas",        "Kansas",        2));
-    /// assert_eq!(true,  filter("sa", &"Mesa",          "Mesa",          3));
-    /// assert_eq!(false, filter("sa", &"Phoenix",       "Phoenix",       4));
-    /// assert_eq!(false, filter("sa", &"Philadelphia",  "Philadelphia",  5));
-    /// assert_eq!(true,  filter("sa", &"San Antonio",   "San Antonio",   6));
-    /// assert_eq!(true,  filter("sa", &"San Diego",     "San Diego",     7));
-    /// assert_eq!(false, filter("sa", &"Dallas",        "Dallas",        8));
-    /// assert_eq!(true,  filter("sa", &"San Francisco", "San Francisco", 9));
-    /// assert_eq!(false, filter("sa", &"Austin",        "Austin",       10));
-    /// assert_eq!(false, filter("sa", &"Jacksonville",  "Jacksonville", 11));
-    /// assert_eq!(true,  filter("sa", &"San Jose",      "San Jose",     12));
+    /// let scorer = MultiSelect::<&str>::DEFAULT_SCORER;
+    /// assert_eq!(None,     scorer("sa", &"New York",      "New York",      0));
+    /// assert_eq!(Some(49), scorer("sa", &"Sacramento",    "Sacramento",    1));
+    /// assert_eq!(Some(35), scorer("sa", &"Kansas",        "Kansas",        2));
+    /// assert_eq!(Some(35), scorer("sa", &"Mesa",          "Mesa",          3));
+    /// assert_eq!(None,     scorer("sa", &"Phoenix",       "Phoenix",       4));
+    /// assert_eq!(None,     scorer("sa", &"Philadelphia",  "Philadelphia",  5));
+    /// assert_eq!(Some(49), scorer("sa", &"San Antonio",   "San Antonio",   6));
+    /// assert_eq!(Some(49), scorer("sa", &"San Diego",     "San Diego",     7));
+    /// assert_eq!(None,     scorer("sa", &"Dallas",        "Dallas",        8));
+    /// assert_eq!(Some(49), scorer("sa", &"San Francisco", "San Francisco", 9));
+    /// assert_eq!(None,     scorer("sa", &"Austin",        "Austin",        10));
+    /// assert_eq!(None,     scorer("sa", &"Jacksonville",  "Jacksonville",  11));
+    /// assert_eq!(Some(49), scorer("sa", &"San Jose",      "San Jose",      12));
     /// ```
-    pub const DEFAULT_FILTER: Filter<'a, T> = &|filter, _, string_value, _| -> bool {
-        let filter = filter.to_lowercase();
+    #[cfg(feature = "fuzzy")]
+    pub const DEFAULT_SCORER: Scorer<'a, T> =
+        &|input, _option, string_value, _idx| -> Option<i64> {
+            DEFAULT_MATCHER.fuzzy_match(string_value, input)
+        };
 
-        string_value.to_lowercase().contains(&filter)
-    };
+    #[cfg(not(feature = "fuzzy"))]
+    pub const DEFAULT_SCORER: Scorer<'a, T> =
+        &|input, _option, string_value, _idx| -> Option<i64> {
+            let filter = input.to_lowercase();
+            match string_value.to_lowercase().contains(&filter) {
+                true => Some(0),
+                false => None,
+            }
+        };
 
     /// Default page size, equal to the global default page size [config::DEFAULT_PAGE_SIZE]
     pub const DEFAULT_PAGE_SIZE: usize = crate::config::DEFAULT_PAGE_SIZE;
@@ -171,6 +196,10 @@ where
 
     /// Default starting cursor index.
     pub const DEFAULT_STARTING_CURSOR: usize = 0;
+
+    /// Default cursor behaviour on filter input change.
+    /// Defaults to true.
+    pub const DEFAULT_RESET_CURSOR: bool = true;
 
     /// Default behavior of keeping or cleaning the current filter value.
     pub const DEFAULT_KEEP_FILTER: bool = true;
@@ -189,8 +218,10 @@ where
             page_size: Self::DEFAULT_PAGE_SIZE,
             vim_mode: Self::DEFAULT_VIM_MODE,
             starting_cursor: Self::DEFAULT_STARTING_CURSOR,
+            starting_filter_input: None,
+            reset_cursor: Self::DEFAULT_RESET_CURSOR,
             keep_filter: Self::DEFAULT_KEEP_FILTER,
-            filter: Self::DEFAULT_FILTER,
+            scorer: Self::DEFAULT_SCORER,
             formatter: Self::DEFAULT_FORMATTER,
             validator: None,
             render_config: get_configuration(),
@@ -227,9 +258,9 @@ where
         self
     }
 
-    /// Sets the filter function.
-    pub fn with_filter(mut self, filter: Filter<'a, T>) -> Self {
-        self.filter = filter;
+    /// Sets the scoring function.
+    pub fn with_scorer(mut self, scorer: Scorer<'a, T>) -> Self {
+        self.scorer = scorer;
         self
     }
 
@@ -273,6 +304,20 @@ where
     /// Sets the starting cursor index.
     pub fn with_starting_cursor(mut self, starting_cursor: usize) -> Self {
         self.starting_cursor = starting_cursor;
+        self
+    }
+
+    /// Sets the starting filter input
+    pub fn with_starting_filter_input(mut self, starting_filter_input: &'a str) -> Self {
+        self.starting_filter_input = Some(starting_filter_input);
+        self
+    }
+
+    /// Sets the reset_cursor behaviour.
+    /// Will reset cursor to first option on filter input change.
+    /// Defaults to true.
+    pub fn with_reset_cursor(mut self, reset_cursor: bool) -> Self {
+        self.reset_cursor = reset_cursor;
         self
     }