+++ inquire/src/prompts/select/mod.rs
@@ -6,7 +6,6 @@ mod prompt;
 mod test;
 
 pub use action::*;
-
 use std::fmt::Display;
 
 use crate::{
@@ -16,12 +15,18 @@ use crate::{
     list_option::ListOption,
     prompts::prompt::Prompt,
     terminal::get_default_terminal,
-    type_aliases::Filter,
+    type_aliases::Scorer,
     ui::{Backend, RenderConfig, SelectBackend},
 };
 
 use self::prompt::SelectPrompt;
 
+#[cfg(feature = "fuzzy")]
+use fuzzy_matcher::{skim::SkimMatcherV2, FuzzyMatcher};
+#[cfg(feature = "fuzzy")]
+use once_cell::sync::Lazy;
+#[cfg(feature = "fuzzy")]
+static DEFAULT_MATCHER: Lazy<SkimMatcherV2> = Lazy::new(|| SkimMatcherV2::default().ignore_case());
 /// Prompt suitable for when you need the user to select one option among many.
 ///
 /// The user can select and submit the current highlighted option by pressing enter.
@@ -38,12 +43,13 @@ use self::prompt::SelectPrompt;
 /// - **Prompt message**: Required when creating the prompt.
 /// - **Options list**: Options displayed to the user. Must be **non-empty**.
 /// - **Starting cursor**: Index of the cursor when the prompt is first rendered. Default is 0 (first option). If the index is out-of-range of the option list, the prompt will fail with an [`InquireError::InvalidConfiguration`] error.
+/// - **Starting filter input**: Sets the initial value of the filter section of the prompt.
 /// - **Help message**: Message displayed at the line below the prompt.
 /// - **Formatter**: Custom formatter in case you need to pre-process the user input before showing it as the final answer.
 ///   - Prints the selected option string value by default.
 /// - **Page size**: Number of options displayed at once, 7 by default.
 /// - **Display option indexes**: On long lists, it might be helpful to display the indexes of the options to the user. Via the `RenderConfig`, you can set the display mode of the indexes as a prefix of an option. The default configuration is `None`, to not render any index when displaying the options.
-/// - **Filter function**: Function that defines if an option is displayed or not based on the current filter input.
+/// - **Scorer function**: Function that defines the order of options and if displayed as all.
 ///
 /// # Example
 ///
@@ -84,9 +90,16 @@ pub struct Select<'a, T> {
     /// Starting cursor index of the selection.
     pub starting_cursor: usize,
 
-    /// Function called with the current user input to filter the provided
+    /// Starting filter input
+    pub starting_filter_input: Option<&'a str>,
+
+    /// Reset cursor position to first option on filter input change.
+    /// Defaults to true.
+    pub reset_cursor: bool,
+
+    /// Function called with the current user input to score the provided
     /// options.
-    pub filter: Filter<'a, T>,
+    pub scorer: Scorer<'a, T>,
 
     /// Function that formats the user input and presents it to the user as the final rendering of the prompt.
     pub formatter: OptionFormatter<'a, T>,
@@ -121,34 +134,44 @@ where
     /// ```
     pub const DEFAULT_FORMATTER: OptionFormatter<'a, T> = &|ans| ans.to_string();
 
-    /// Default filter function, which checks if the current filter value is a substring of the option value.
-    /// If it is, the option is displayed.
+    /// Default scoring function, which will create a score for the current option using the input value.
+    /// The return will be sorted in Descending order, leaving options with None as a score.
     ///
     /// # Examples
     ///
     /// ```
     /// use inquire::Select;
     ///
-    /// let filter = Select::<&str>::DEFAULT_FILTER;
-    /// assert_eq!(false, filter("sa", &"New York",      "New York",      0));
-    /// assert_eq!(true,  filter("sa", &"Sacramento",    "Sacramento",    1));
-    /// assert_eq!(true,  filter("sa", &"Kansas",        "Kansas",        2));
-    /// assert_eq!(true,  filter("sa", &"Mesa",          "Mesa",          3));
-    /// assert_eq!(false, filter("sa", &"Phoenix",       "Phoenix",       4));
-    /// assert_eq!(false, filter("sa", &"Philadelphia",  "Philadelphia",  5));
-    /// assert_eq!(true,  filter("sa", &"San Antonio",   "San Antonio",   6));
-    /// assert_eq!(true,  filter("sa", &"San Diego",     "San Diego",     7));
-    /// assert_eq!(false, filter("sa", &"Dallas",        "Dallas",        8));
-    /// assert_eq!(true,  filter("sa", &"San Francisco", "San Francisco", 9));
-    /// assert_eq!(false, filter("sa", &"Austin",        "Austin",       10));
-    /// assert_eq!(false, filter("sa", &"Jacksonville",  "Jacksonville", 11));
-    /// assert_eq!(true,  filter("sa", &"San Jose",      "San Jose",     12));
+    /// let scorer = Select::<&str>::DEFAULT_SCORER;
+    /// assert_eq!(None,     scorer("sa", &"New York",      "New York",      0));
+    /// assert_eq!(Some(49), scorer("sa", &"Sacramento",    "Sacramento",    1));
+    /// assert_eq!(Some(35), scorer("sa", &"Kansas",        "Kansas",        2));
+    /// assert_eq!(Some(35), scorer("sa", &"Mesa",          "Mesa",          3));
+    /// assert_eq!(None,     scorer("sa", &"Phoenix",       "Phoenix",       4));
+    /// assert_eq!(None,     scorer("sa", &"Philadelphia",  "Philadelphia",  5));
+    /// assert_eq!(Some(49), scorer("sa", &"San Antonio",   "San Antonio",   6));
+    /// assert_eq!(Some(49), scorer("sa", &"San Diego",     "San Diego",     7));
+    /// assert_eq!(None,     scorer("sa", &"Dallas",        "Dallas",        8));
+    /// assert_eq!(Some(49), scorer("sa", &"San Francisco", "San Francisco", 9));
+    /// assert_eq!(None,     scorer("sa", &"Austin",        "Austin",        10));
+    /// assert_eq!(None,     scorer("sa", &"Jacksonville",  "Jacksonville",  11));
+    /// assert_eq!(Some(49), scorer("sa", &"San Jose",      "San Jose",      12));
     /// ```
-    pub const DEFAULT_FILTER: Filter<'a, T> = &|filter, _, string_value, _| -> bool {
-        let filter = filter.to_lowercase();
+    #[cfg(feature = "fuzzy")]
+    pub const DEFAULT_SCORER: Scorer<'a, T> =
+        &|input, _option, string_value, _idx| -> Option<i64> {
+            DEFAULT_MATCHER.fuzzy_match(string_value, input)
+        };
 
-        string_value.to_lowercase().contains(&filter)
-    };
+    #[cfg(not(feature = "fuzzy"))]
+    pub const DEFAULT_SCORER: Scorer<'a, T> =
+        &|input, _option, string_value, _idx| -> Option<i64> {
+            let filter = input.to_lowercase();
+            match string_value.to_lowercase().contains(&filter) {
+                true => Some(0),
+                false => None,
+            }
+        };
 
     /// Default page size.
     pub const DEFAULT_PAGE_SIZE: usize = crate::config::DEFAULT_PAGE_SIZE;
@@ -158,6 +181,10 @@ where
 
     /// Default starting cursor index.
     pub const DEFAULT_STARTING_CURSOR: usize = 0;
+
+    /// Default cursor behaviour on filter input change.
+    /// Defaults to true.
+    pub const DEFAULT_RESET_CURSOR: bool = true;
 
     /// Default help message.
     pub const DEFAULT_HELP_MESSAGE: Option<&'a str> =
@@ -172,9 +199,11 @@ where
             page_size: Self::DEFAULT_PAGE_SIZE,
             vim_mode: Self::DEFAULT_VIM_MODE,
             starting_cursor: Self::DEFAULT_STARTING_CURSOR,
-            filter: Self::DEFAULT_FILTER,
+            reset_cursor: Self::DEFAULT_RESET_CURSOR,
+            scorer: Self::DEFAULT_SCORER,
             formatter: Self::DEFAULT_FORMATTER,
             render_config: get_configuration(),
+            starting_filter_input: None,
         }
     }
 
@@ -202,9 +231,9 @@ where
         self
     }
 
-    /// Sets the filter function.
-    pub fn with_filter(mut self, filter: Filter<'a, T>) -> Self {
-        self.filter = filter;
+    /// Sets the scoring function.
+    pub fn with_scorer(mut self, scorer: Scorer<'a, T>) -> Self {
+        self.scorer = scorer;
         self
     }
 
@@ -217,6 +246,20 @@ where
     /// Sets the starting cursor index.
     pub fn with_starting_cursor(mut self, starting_cursor: usize) -> Self {
         self.starting_cursor = starting_cursor;
+        self
+    }
+
+    /// Sets the starting filter input
+    pub fn with_starting_filter_input(mut self, starting_filter_input: &'a str) -> Self {
+        self.starting_filter_input = Some(starting_filter_input);
+        self
+    }
+
+    /// Sets the reset_cursor behaviour.
+    /// Will reset cursor to first option on filter input change.
+    /// Defaults to true.
+    pub fn with_reset_cursor(mut self, reset_cursor: bool) -> Self {
+        self.reset_cursor = reset_cursor;
         self
     }